Ensure SIGTERM is passed to PID 1 on OOM error
Running the Sidekiq (or any other) process as a child of 'run' caused
the process to NOT get any signal when it was killed from out of memory
errors.

By passing 'exec' when we call it, we substitut the PID 1 of the
container from 's2i/bin/run' to the actual process running, which is
what we want to get the SIGTERM signal.